kernel: Fix console_print() unsafe memory access
Rewrite console_print() to use copy_from_user(), which fails gracefully
if the memory pointer provided by user space is not mapped to process
address space.

Fixes #17

+++ include/kernel/syscall.h
@@ -7,6 +7,6 @@ enum {
 	SYS_console_print	= 3,
 };
 
-int syscall(int nr, ...);
+long syscall(int nr, ...);
 
 #endif
